Hide language toggle on admin page (#128)
Hide the language bar when logged in. Only English will show up. When you hit the language toggle by accident when logged in it navigates you away from the admin page and it's confusing to have the language toggle when admin is English-only. If we ever translate admin into French we may change this but right now it's English only and we definitely aren't translating admin into dozens of languages because there's no point.