Send multiple posts for each vaccine type checked (#185)
Changes the vaccine selection from a dropdown to a list of checkboxes to
allow for multiple vaccines to be added at once. This allows a
pharmacist to add a new set of rapid appointment availability for
several different vaccine types without having to submit the form
multipler times.

We accomplish it by sending off one request for each vaccine type. If
none are selected (or if the "Not Sure" option is selected) then we'll
send vaccine type "1". Otherwise each request will use the numeric type
of the named vaccine.