Use the callback format for toggle state updates (#188)
When toggling a value from true to false, we'd prefer to use the
callback approach. This ensures that any actions that get dispatched
will operate on the true value of the state, and not what happens to
currently be set in the scope of the component. Niche, but good
practice.